[Train] Clean up `ray.train` package (#25566)
