+++ doc/source/train/api.rst
@@ -109,32 +109,6 @@ TorchTensorboardProfilerCallback
 
 .. autoclass:: ray.train.callbacks.TorchTensorboardProfilerCallback
 
-ResultsPreprocessors
-~~~~~~~~~~~~~~~~~~~~
-
-.. _train-api-results-preprocessor:
-
-ResultsPreprocessor
-+++++++++++++++++++
-
-.. autoclass:: ray.train.callbacks.results_preprocessors.ResultsPreprocessor
-    :members:
-
-SequentialResultsPreprocessor
-+++++++++++++++++++++++++++++++
-
-.. autoclass:: ray.train.callbacks.results_preprocessors.SequentialResultsPreprocessor
-
-IndexedResultsPreprocessor
-+++++++++++++++++++++++++++++++
-
-.. autoclass:: ray.train.callbacks.results_preprocessors.IndexedResultsPreprocessor
-
-ExcludedKeysResultsPreprocessor
-+++++++++++++++++++++++++++++++
-
-.. autoclass:: ray.train.callbacks.results_preprocessors.ExcludedKeysResultsPreprocessor
-
 Checkpointing
 -------------
