+++ python/ray/train/__init__.py
@@ -1,6 +1,6 @@
 from ray.train.backend import BackendConfig
 from ray.train.callbacks import TrainingCallback
-from ray.train.session import (
+from ray.train.train_loop_utils import (
     get_dataset_shard,
     local_rank,
     load_checkpoint,
