+++ python/ray/train/_internal/results_preprocessors/preprocessor.py
@@ -1,10 +1,7 @@
 import abc
 from typing import List, Dict
 
-from ray.util.annotations import DeveloperAPI
 
-
-@DeveloperAPI
 class ResultsPreprocessor(abc.ABC):
     """Abstract class for preprocessing Train results."""
 
@@ -27,7 +24,6 @@ class ResultsPreprocessor(abc.ABC):
         return results
 
 
-@DeveloperAPI
 class SequentialResultsPreprocessor(ResultsPreprocessor):
     """A processor that sequentially runs a series of preprocessing steps.
