+++ python/ray/train/callbacks/print.py
@@ -2,8 +2,10 @@ import json
 from typing import Dict, List
 
 from ray.train.callbacks import TrainingCallback
+from ray.util.annotations import PublicAPI
 
 
+@PublicAPI(stability="beta")
 class PrintCallback(TrainingCallback):
     """A callback that prints training results to STDOUT.
