+++ python/ray/train/error.py
@@ -1,0 +1,6 @@
+from ray.util.annotations import PublicAPI
+
+
+@PublicAPI(stability="beta")
+class SessionMisuseError(Exception):
+    """Indicates a method or function was used outside of a session."""
