+++ python/ray/train/horovod/__init__.py
@@ -1,0 +1,12 @@
+try:
+    import horovod  # noqa: F401
+except ModuleNotFoundError:
+    raise ModuleNotFoundError(
+        "Horovod isn't installed. To install Horovod with PyTorch support, run 'pip "
+        "install 'horovod[pytorch]''. To install Horovod with TensorFlow support, "
+        "run 'pip install 'horovod[tensorflow]''."
+    )
+
+from ray.train.horovod.config import HorovodConfig
+
+__all__ = ["HorovodConfig"]
