+++ python/ray/train/horovod/config.py
@@ -1,27 +1,18 @@
-import logging
+from typing import Optional, Set
+
 import os
 from dataclasses import dataclass
-from typing import Optional, Set
 
 import ray
 from ray.train.backend import BackendConfig, Backend
-from ray.train.utils import update_env_vars
-from ray.train.worker_group import WorkerGroup, Worker
+from ray.train._internal.utils import update_env_vars
+from ray.train._internal.worker_group import WorkerGroup, Worker
 
-try:
-    from horovod.ray.runner import Coordinator
-    from horovod.ray.utils import detect_nics, nics_to_env_var
-    from horovod.runner.common.util import secret, timeout
-except ModuleNotFoundError:
-    raise ModuleNotFoundError(
-        "Horovod isn't installed. To install Horovod with PyTorch support, run 'pip "
-        "install 'horovod[pytorch]''. To install Horovod with TensorFlow support, "
-        "run 'pip install 'horovod[tensorflow]''."
-    )
+from horovod.ray.runner import Coordinator
+from horovod.ray.utils import detect_nics, nics_to_env_var
+from horovod.runner.common.util import secret, timeout
 
 from ray.util import PublicAPI
-
-logger = logging.getLogger(__name__)
 
 
 @PublicAPI(stability="beta")
@@ -76,40 +67,10 @@ class HorovodConfig(BackendConfig):
 
     @property
     def backend_cls(self):
-        return HorovodBackend
+        return _HorovodBackend
 
 
-def init_env_vars(world_rank: int, world_size: int, node_id: str):
-    """Initialize Horovod environment variables."""
-    os.environ["HOROVOD_HOSTNAME"] = node_id
-    os.environ["HOROVOD_RANK"] = str(world_rank)
-    os.environ["HOROVOD_SIZE"] = str(world_size)
-
-
-# TODO(tgaddair): temporary workaround for Horovod's worker discovery logic,
-#  which requires passing in an extra parameter as part of the RayExecutor
-#  API. This will be removed in the future as we migrate more of the
-#  RayExecutor utils into Ray Train.
-#  See: https://github.com/horovod/horovod/blob/v0.23.0/horovod/ray/driver_service.py#L9 # noqa: E501
-@dataclass
-class HorovodWorkerWrapper:
-    w: Worker
-
-    @property
-    def execute(self):
-        w = self.w
-
-        class ExecuteHandle:
-            def remote(self, func, *args, **kwargs):
-                _ = None
-                return w.actor._BaseWorkerMixin__execute.remote(
-                    func, _, *args, **kwargs
-                )
-
-        return ExecuteHandle()
-
-
-class HorovodBackend(Backend):
+class _HorovodBackend(Backend):
     share_cuda_visible_devices: bool = True
 
     def on_start(self, worker_group: WorkerGroup, backend_config: HorovodConfig):
@@ -122,7 +83,7 @@ class HorovodBackend(Backend):
             worker_node_id = worker_group.workers[rank].metadata.node_id
             setup_futures.append(
                 worker_group.execute_single_async(
-                    rank, init_env_vars, rank, len(worker_group), worker_node_id
+                    rank, _init_env_vars, rank, len(worker_group), worker_node_id
                 )
             )
         ray.get(setup_futures)
@@ -154,7 +115,7 @@ class HorovodBackend(Backend):
         # Get one worker from each host/node.
         node_worker_indexes = [node_ids.index(node_id) for node_id in set(node_ids)]
         node_workers = [
-            HorovodWorkerWrapper(worker_group.workers[worker_index])
+            _HorovodWorkerWrapper(worker_group.workers[worker_index])
             for worker_index in node_worker_indexes
         ]
         assert len(node_workers) == len(self.coordinator.hostnames)
@@ -167,3 +128,33 @@ class HorovodBackend(Backend):
         coordinator_envs.update(nics_to_env_var(nics))
 
         worker_group.execute(update_env_vars, coordinator_envs)
+
+
+def _init_env_vars(world_rank: int, world_size: int, node_id: str):
+    """Initialize Horovod environment variables."""
+    os.environ["HOROVOD_HOSTNAME"] = node_id
+    os.environ["HOROVOD_RANK"] = str(world_rank)
+    os.environ["HOROVOD_SIZE"] = str(world_size)
+
+
+# TODO(tgaddair): temporary workaround for Horovod's worker discovery logic,
+#  which requires passing in an extra parameter as part of the RayExecutor
+#  API. This will be removed in the future as we migrate more of the
+#  RayExecutor utils into Ray Train.
+#  See: https://github.com/horovod/horovod/blob/v0.23.0/horovod/ray/driver_service.py#L9 # noqa: E501
+@dataclass
+class _HorovodWorkerWrapper:
+    w: Worker
+
+    @property
+    def execute(self):
+        w = self.w
+
+        class ExecuteHandle:
+            def remote(self, func, *args, **kwargs):
+                _ = None
+                return w.actor._BaseWorkerMixin__execute.remote(
+                    func, _, *args, **kwargs
+                )
+
+        return ExecuteHandle()
