+++ python/ray/train/tensorflow/__init__.py
@@ -1,0 +1,12 @@
+try:
+    import tensorflow as tf  # noqa: F401
+except ModuleNotFoundError:
+    raise ModuleNotFoundError(
+        "TensorFlow isn't installed. To install TensorFlow, run 'pip install "
+        "tensorflow'."
+    )
+
+from ray.train.tensorflow.config import TensorflowConfig
+from ray.train.tensorflow.train_loop_utils import prepare_dataset_shard
+
+__all__ = ["TensorflowConfig", "prepare_dataset_shard"]
