+++ python/ray/train/tensorflow/config.py
@@ -6,17 +6,10 @@ from typing import List
 
 import ray
 from ray.train.backend import BackendConfig, Backend
-from ray.train.utils import get_address_and_port
-from ray.train.worker_group import WorkerGroup
+from ray.train._internal.utils import get_address_and_port
+from ray.train._internal.worker_group import WorkerGroup
 from ray.util import PublicAPI
 
-try:
-    import tensorflow as tf
-except ModuleNotFoundError:
-    raise ModuleNotFoundError(
-        "TensorFlow isn't installed. To install TensorFlow, run 'pip install "
-        "tensorflow'."
-    )
 
 logger = logging.getLogger(__name__)
 
@@ -26,10 +19,10 @@ logger = logging.getLogger(__name__)
 class TensorflowConfig(BackendConfig):
     @property
     def backend_cls(self):
-        return TensorflowBackend
+        return _TensorflowBackend
 
 
-def setup_tensorflow_environment(worker_addresses: List[str], index: int):
+def _setup_tensorflow_environment(worker_addresses: List[str], index: int):
     """Set up distributed Tensorflow training information.
 
     This function should be called on each worker.
@@ -45,7 +38,7 @@ def setup_tensorflow_environment(worker_addresses: List[str], index: int):
     os.environ["TF_CONFIG"] = json.dumps(tf_config)
 
 
-class TensorflowBackend(Backend):
+class _TensorflowBackend(Backend):
     def on_start(self, worker_group: WorkerGroup, backend_config: TensorflowConfig):
         # Compute URL for initializing distributed setup.
         def get_url():
@@ -59,28 +52,7 @@ class TensorflowBackend(Backend):
         for i in range(len(worker_group)):
             setup_futures.append(
                 worker_group.execute_single_async(
-                    i, setup_tensorflow_environment, worker_addresses=urls, index=i
+                    i, _setup_tensorflow_environment, worker_addresses=urls, index=i
                 )
             )
         ray.get(setup_futures)
-
-
-@PublicAPI(stability="beta")
-def prepare_dataset_shard(tf_dataset_shard: tf.data.Dataset):
-    """A utility function that disables Tensorflow autosharding.
-
-    This should be used on a  TensorFlow ``Dataset`` created by calling ``to_tf()``
-    on a ``ray.data.Dataset`` returned by ``ray.train.get_dataset_shard()`` since
-    the dataset has already been sharded across the workers.
-
-    Args:
-        tf_dataset_shard (tf.data.Dataset): A TensorFlow Dataset.
-
-    Returns:
-        A TensorFlow Dataset with autosharding turned off.
-    """
-    options = tf.data.Options()
-    options.experimental_distribute.auto_shard_policy = (
-        tf.data.experimental.AutoShardPolicy.OFF
-    )
-    return tf_dataset_shard.with_options(options)
