+++ python/ray/train/tests/test_backend.py
@@ -7,21 +7,21 @@ import pytest
 import ray
 import ray.train as train
 from ray.cluster_utils import Cluster
-from ray.train.backend import (
-    Backend,
+from ray.train.backend import Backend, BackendConfig
+from ray.train._internal.backend_executor import (
+    BackendExecutor,
     InactiveWorkerGroupError,
     TrainBackendError,
     TrainingWorkerError,
 )
-from ray.train.backend import BackendConfig, BackendExecutor
-from ray.train.impl.dataset_spec import _RayDatasetSpec
+from ray.train._internal.dataset_spec import RayDatasetSpec
 from ray.train.tensorflow import TensorflowConfig
 from ray.train.torch import TorchConfig
 from ray.train.constants import (
     ENABLE_SHARE_CUDA_VISIBLE_DEVICES_ENV,
     TRAIN_ENABLE_WORKER_SPREAD_ENV,
 )
-from ray.train.worker_group import WorkerGroup
+from ray.train._internal.worker_group import WorkerGroup
 from ray.util.placement_group import get_current_placement_group
 
 # Trigger pytest hook to automatically zip test cluster logs to archive dir on failure
@@ -103,7 +103,7 @@ class TestBackend(Backend):
         pass
 
 
-EMPTY_RAY_DATASET_SPEC = _RayDatasetSpec(dataset_or_dict=None)
+EMPTY_RAY_DATASET_SPEC = RayDatasetSpec(dataset_or_dict=None)
 
 
 def test_start(ray_start_2_cpus):
