+++ python/ray/train/tests/test_trainer.py
@@ -11,14 +11,15 @@ import ray
 import ray.train as train
 from ray._private.test_utils import wait_for_condition
 from ray.train import Trainer, CheckpointStrategy
-from ray.train.backend import BackendConfig, Backend, BackendExecutor
+from ray.train.backend import BackendConfig, Backend
 from ray.train.constants import TRAIN_ENABLE_WORKER_SPREAD_ENV
 from ray.train.torch import TorchConfig
 from ray.train.tensorflow import TensorflowConfig
 
 from ray.train.horovod import HorovodConfig
 from ray.train.callbacks.callback import TrainingCallback
-from ray.train.worker_group import WorkerGroup
+from ray.train._internal.worker_group import WorkerGroup
+from ray.train._internal.backend_executor import BackendExecutor
 
 
 @pytest.fixture
